+++ Cargo.toml
@@ -13,11 +13,14 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.1.8", features = ["cargo", "derive", "help", "error-context", "std", "usage", "wrap_help"], default_features = false }
 crossterm = "0.26.0"
+rustls-pemfile = "1.0.2"
 tokio = { version = "1.14.0", features = ["full"] }
+tokio-rustls = { version = "0.23.4" }
 tracing = "0.1.11"
 tracing-subscriber = "0.3.16"
 tui = "0.19.0"
 urlencoding = "2.1.2"
+webpki-roots = "0.22.6"
 
 [lib]
 name = "mtop"
