+++ tools/tls-connect.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -o errexit
+
+exec openssl s_client -CAfile memcached-ca-cert.pem -cert memcached-client-cert.pem -key memcached-client-key.pem -connect localhost:11211
